[#175265955] add paymento method from bpd screen
